pyproject.toml
[tool.poetry]
name = "weather-dashboard"
version = "0.1.0"
description = ""
authors = ["Fakile Temitope Oluwaseyi <fakiletemitope@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
boto3 = "^1.35.92"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
